django_web/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.2
numpy==1.21.1
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
paho-mqtt




